requirements.txt
newsapi-python
pymysql
sys
hashlib